Add support for sigmoid animation curve
The curve is tuned so that the majority of the moves take place (smoothly)
between 9am and 6pm, with the early opening and late endgame slowly spread
out over the morning and evening. This makes for a much more exciting progression
than turning your computer on in the morning with the game already 40% completed.